requirements.txt
PyQt5==5.15.11
PyQt5_sip==12.15.0
python_docx==1.1.2
Requests==2.32.3
